requirements.txt
ansi2html==1.9.2
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blinker==1.9.0
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==2.5.0
dash==4.0.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==6.0.3
django-plotly-dash==2.5.0
dpd_components==0.2.0
Flask==2.1.3
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
narwhals==2.17.0
nest-asyncio==1.6.0
packaging==26.0
plotly==6.6.0
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==25.3.0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
retrying==1.4.2
service-identity==24.2.0
setuptools==82.0.0
sqlparse==0.5.5
tenacity==9.1.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
Werkzeug==2.0.3
zipp==3.23.0
zope.interface==8.2
